requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
PyYAML>=6.0.0
APScheduler>=3.10.0
tenacity>=8.2.0
python-dateutil>=2.9.0
